move common functions to common.js in fixtures
